Issue #68: Split significance tests for 300-LOC compliance
- Extract shared LaTeX parsers and tolerance constants to analysis/tests/_dynreg_tex.py (277 LOC)
- Append dynamic-regression fixtures to analysis/tests/conftest.py (session-scoped baseline_coefs, extended_coefs, baseline_details, baseline_gof)
- Create analysis/tests/test_dynamic_regression_parity.py (249 LOC) with baseline coef/SE/GoF parity tests and cross-table extended/baseline comparisons
- Trim analysis/tests/test_dynamic_regression_significance.py to 72 LOC (extended structural tests + 21 Chat/Facial coefficient pins only)
- Update wiki/methods/dynamic-regression.md to reflect split (12 parity tests + 2 extended structure tests)
- All 143/143 dynamic regression tests pass across 6 test files
