+++ analysis/tests/test_dynamic_regression_significance.py
@@ -1,146 +1,23 @@
 """
-Coefficient parity tests for dynamic_regression.R against Table DP1 reference.
+Structural + pinning tests for dynamic_regression_extended.tex.
 
-Parses output/tables/dynamic_regression_baseline.tex (4 cols) and
-output/tables/dynamic_regression_extended.tex (12 cols). Verifies that the
-baseline coefficients, standard errors, significance stars, and goodness-of-fit
-rows match analysis/issues/issue_68_table_dp1_reference.txt within tolerance.
+Verifies the extended table exposes 12 data columns and that the Chat/Facial
+coefficients added by +Chat and +Chat+Facial specs match their pinned values.
+Coefficient-parity against Stata Table DP1 lives in test_dynamic_regression_parity.py.
 
-Author: Claude Code (test-writer)
-Date: 2026-04-19
+Author: Claude Code
+Date: 2026-04-20
 """
 
-import re
-import subprocess
+import sys
 from pathlib import Path
 
-import pytest
+sys.path.insert(0, str(Path(__file__).parent))
+from _dynreg_tex import check_coef  # noqa: E402
 
-# FILE PATHS
-REPO_ROOT = Path(__file__).parent.parent.parent
-ANALYSIS_DIR = REPO_ROOT / "analysis"
-R_SCRIPT = ANALYSIS_DIR / "analysis" / "dynamic_regression.R"
-INPUT_CSV = ANALYSIS_DIR / "datastore" / "derived" / "dynamic_regression_panel.csv"
-BASELINE_TEX = ANALYSIS_DIR / "output" / "tables" / "dynamic_regression_baseline.tex"
-EXTENDED_TEX = ANALYSIS_DIR / "output" / "tables" / "dynamic_regression_extended.tex"
-
-# 0.005 matches Stata's 3-decimal log precision; any drift >0.005 indicates a
-# genuine spec mismatch. Per-coefficient overrides handle legitimate numerical
-# drift on large-magnitude intercepts.
-TOLERANCE = 0.005
-TOLERANCE_OVERRIDES = {
-    # Round 1 intercepts (~|12|) drift ~0.006 from Stata due to GMM two-step
-    # optimizer differences on large-scale constants. Relative drift <0.05%.
-    "Round 1": 0.01,
-    # "Above max peer" lands at the 0.005 boundary (Stata 0.064 vs R 0.059);
-    # the drift equals Stata's 3-decimal rounding precision so an epsilon is
-    # needed to avoid a false-positive from float comparison.
-    "Above max peer$_{t-1}$": 0.006,
-}
-SE_TOLERANCE = 0.005
-# GoF p-values are 3-digit rounded; allow 0.002 rounding slack.
-GOF_TOLERANCE = 0.002
-
-# Column order in dynamic_regression_baseline.tex (0-indexed):
-#   0: T1 (mean), 1: T2 (mean), 2: T1 (min/med/max), 3: T2 (min/med/max)
-REFERENCE_COLUMN_INDEX = {
-    "T1_mean":       0,
-    "T2_mean":       1,
-    "T1_minmedmax":  2,
-    "T2_minmedmax":  3,
-}
-
-T1_MEAN_REFERENCE = {
-    "Contribution$_{t-1}$":       -0.069,
-    "Contribution$_{t-2}$":       -0.181,
-    "Above peer mean$_{t-1}$": -0.406,
-    "Below peer mean$_{t-1}$":  0.268,
-    "Round 1":                   -12.715,
-}
-
-T2_MEAN_REFERENCE = {
-    "Contribution$_{t-1}$":        0.087,
-    "Contribution$_{t-2}$":       -0.242,
-    "Above peer mean$_{t-1}$": -0.263,
-    "Below peer mean$_{t-1}$":  0.553,
-    "Round 1":                   -5.591,
-}
-
-T1_MINMEDMAX_REFERENCE = {
-    "Contribution$_{t-1}$":  -0.056,
-    "Contribution$_{t-2}$":  -0.177,
-    "Above max peer$_{t-1}$":    0.064,
-    "Below max peer$_{t-1}$":    0.071,
-    "Above median peer$_{t-1}$":   -0.179,
-    "Below median peer$_{t-1}$":    0.201,
-    "Above min peer$_{t-1}$":   -0.160,
-    "Below min peer$_{t-1}$":   -0.016,
-    "Round 1":              -12.811,
-}
-
-# Stata reference for T2 min/med/max is not in the reference .txt (issue_68
-# Stata do-file did not export this model). Values pinned from current R
-# output, already verified by the team to match Stata within 0.01.
-T2_MINMEDMAX_REFERENCE = {
-    "Contribution$_{t-1}$":    0.473,
-    "Contribution$_{t-2}$":   -0.268,
-    "Above max peer$_{t-1}$":     0.056,
-    "Below max peer$_{t-1}$":     0.819,
-    "Above median peer$_{t-1}$":    -0.069,
-    "Below median peer$_{t-1}$":    -0.089,
-    "Above min peer$_{t-1}$":    -0.117,
-    "Below min peer$_{t-1}$":     0.156,
-    "Round 1":                -5.221,
-}
-
-# Standard errors for T1 mean column from reference file (Stata SEs).
-T1_MEAN_SE_REFERENCE = {
-    "Contribution$_{t-1}$":       0.027,
-    "Contribution$_{t-2}$":       0.020,
-    "Above peer mean$_{t-1}$": 0.049,
-    "Below peer mean$_{t-1}$": 0.065,
-    "Round 1":                    0.742,
-}
-
-# Significance stars for T1 mean from reference file: *** = p<0.01, ** = p<0.05,
-# * = p<0.1, "" = not significant.
-T1_MEAN_STARS_REFERENCE = {
-    "Contribution$_{t-1}$":       "***",
-    "Contribution$_{t-2}$":       "***",
-    "Above peer mean$_{t-1}$": "***",
-    "Below peer mean$_{t-1}$": "***",
-    "Round 1":                    "***",
-}
-
-# GoF reference values: Observations from issue_68_table_dp1_reference.txt,
-# p-values from current .tex output (cross-checked against Stata within 0.01).
-T1_MEAN_GOF_REFERENCE = {
-    "Observations":           1520,
-    "AR(1) p-value":          0.000,
-    "AR(2) p-value":          0.183,
-    "Sargan p-value":         0.231,
-    "Peer mean pair sum = 0 (p)":          0.137,
-}
-T2_MEAN_GOF_REFERENCE = {
-    "Observations":           1520,
-    "AR(1) p-value":          0.000,
-    "AR(2) p-value":          0.309,
-    "Sargan p-value":         0.252,
-    "Peer mean pair sum = 0 (p)":          0.006,
-}
-T1_MINMEDMAX_GOF_REFERENCE = {
-    "Observations":               1520,
-    "AR(1) p-value":              0.000,
-    "AR(2) p-value":              0.193,
-    "Sargan p-value":             0.234,
-    "Max peer pair sum = 0 (p)":      0.206,
-    "Median peer pair sum = 0 (p)":      0.827,
-    "Min peer pair sum = 0 (p)":      0.059,
-}
-
-# Extended-table pins for chat/facial coefficients (cols 1,2,4,5,7,8,10,11 —
-# i.e. +Chat and +Chat+Facial specs for each baseline). Values from current
-# output, locking in the "correct" numbers against future regressions.
+# Chat/Facial coefficient pins for extended cols 1,2,4,5,7,8,10,11 (i.e. +Chat
+# and +Chat+Facial specs for each baseline). Values frozen from current R
+# output so a regression changes these visibly.
 EXTENDED_CHAT_FACIAL_PINS = {
     # T1 mean +Chat (col 1), +Chat+Facial (col 2)
     ("Word Count", 1):             0.079,
@@ -169,445 +46,11 @@ EXTENDED_CHAT_FACIAL_PINS = {
 }
 
 
-# =====
-# Main function (FIRST - shows high-level flow)
-# =====
-def main():
-    """Manual invocation: parse both .tex files and print the parsed rows."""
-    baseline = parse_tex_coefficients(BASELINE_TEX, num_data_cols=4)
-    extended = parse_tex_coefficients(EXTENDED_TEX, num_data_cols=12)
-    print("Baseline:")
-    for label, row in baseline.items():
-        print(f"  {label}: {row}")
-    print(f"Extended: {len(extended)} rows, first = {next(iter(extended.items()))}")
-
-
-# =====
-# Fixtures
-# =====
-@pytest.fixture(scope="module")
-def baseline_coefs():
-    """Ensure baseline .tex exists and parse into {label: [coefs per column]}."""
-    ensure_tex_outputs_current()
-    if not BASELINE_TEX.exists():
-        raise FileNotFoundError(
-            f"Missing baseline table: {BASELINE_TEX}. "
-            f"Run: cd analysis && Rscript analysis/dynamic_regression.R"
-        )
-    return parse_tex_coefficients(BASELINE_TEX, num_data_cols=4)
-
-
-@pytest.fixture(scope="module")
-def extended_coefs():
-    """Parse the 12-column extended .tex into {label: [coefs per column]}."""
-    ensure_tex_outputs_current()
-    if not EXTENDED_TEX.exists():
-        raise FileNotFoundError(
-            f"Missing extended table: {EXTENDED_TEX}. "
-            f"Run: cd analysis && Rscript analysis/dynamic_regression.R"
-        )
-    return parse_tex_coefficients(EXTENDED_TEX, num_data_cols=12)
-
-
-@pytest.fixture(scope="module")
-def baseline_details():
-    """Parse baseline .tex with SE rows and stars: {label: [(coef, se, stars), ...]}."""
-    ensure_tex_outputs_current()
-    return parse_tex_with_details(BASELINE_TEX, num_data_cols=4)
-
-
-@pytest.fixture(scope="module")
-def baseline_gof():
-    """Parse the GoF rows (Observations, AR/Sargan/Wald p-values) from baseline.tex."""
-    ensure_tex_outputs_current()
-    return parse_tex_gof_rows(BASELINE_TEX, num_data_cols=4)
-
-
-def ensure_tex_outputs_current():
-    """Re-run dynamic_regression.R if either .tex is missing or stale vs sources."""
-    if not R_SCRIPT.exists():
-        raise FileNotFoundError(f"Missing R script: {R_SCRIPT}")
-    if not _tex_outputs_stale():
-        return
-    result = _run_r_script_with_timeout()
-    if result.returncode != 0:
-        raise RuntimeError(
-            f"dynamic_regression.R failed (exit {result.returncode}).\n"
-            f"STDOUT:\n{result.stdout}\nSTDERR:\n{result.stderr}"
-        )
-
-
-def _run_r_script_with_timeout():
-    """Invoke Rscript with 300s timeout; convert TimeoutExpired to actionable msg."""
-    try:
-        return subprocess.run(
-            ["Rscript", "analysis/dynamic_regression.R"],
-            cwd=ANALYSIS_DIR, capture_output=True, text=True, timeout=300,
-        )
-    except subprocess.TimeoutExpired as exc:
-        raise RuntimeError(
-            f"dynamic_regression.R timed out after 300s. "
-            f"Investigate hung process or data issue. cmd={exc.cmd}"
-        ) from exc
-
-
-def _tex_outputs_stale() -> bool:
-    """Return True if any .tex output is missing or older than R_SCRIPT/INPUT_CSV."""
-    if not BASELINE_TEX.exists() or not EXTENDED_TEX.exists():
-        return True
-    source_mtimes = [R_SCRIPT.stat().st_mtime]
-    if INPUT_CSV.exists():
-        source_mtimes.append(INPUT_CSV.stat().st_mtime)
-    newest_source = max(source_mtimes)
-    oldest_tex = min(BASELINE_TEX.stat().st_mtime, EXTENDED_TEX.stat().st_mtime)
-    return newest_source > oldest_tex
-
-
-# =====
-# .tex parsing: coefficient cells
-# =====
-# Matches a numeric coefficient cell with optional stars: $-0.069^{***}$ or $0.268$.
-COEF_CELL_RE = re.compile(r"\$(-?\d+(?:\.\d+)?)(?:\^\{(\*+)\})?\$")
-# Matches a standard error parenthetical cell: $(0.049)$.
-SE_CELL_RE = re.compile(r"\$\((-?\d+(?:\.\d+)?)\)\$")
-
-
-def _is_structural_line(line: str) -> bool:
-    """Return True for non-data tabular lines (rules, multicol, resizebox)."""
-    structural = ("\\toprule", "\\midrule", "\\bottomrule",
-                  "\\cmidrule", "\\multicolumn", "\\resizebox")
-    return any(tok in line for tok in structural)
-
-
-def _iter_tabular_rows(text: str):
-    """Yield raw cell-lists for each tabular data row, in order."""
-    in_tabular = False
-    for raw in text.splitlines():
-        line = raw.rstrip()
-        if "\\begin{tabular}" in line or "\\begin{tabular*}" in line:
-            in_tabular = True
-            continue
-        if "\\end{tabular}" in line or "\\end{tabular*}" in line:
-            break
-        if not in_tabular or "\\" not in line or _is_structural_line(line):
-            continue
-        cells = split_tex_row(line)
-        if cells is not None:
-            yield cells
-
-
-def _parse_tabular_body(text: str, num_data_cols: int) -> dict:
-    """Extract coefficient rows (label -> [floats]) from the tabular body."""
-    rows = {}
-    for cells in _iter_tabular_rows(text):
-        if len(cells) != num_data_cols + 1:
-            continue
-        label = cells[0].strip()
-        if label and label not in rows:
-            rows[label] = [parse_cell(c) for c in cells[1:]]
-    return rows
-
-
-def parse_tex_coefficients(tex_path: Path, num_data_cols: int) -> dict:
-    """Parse dynamic_regression_*.tex -> {coef_label: [float|None, ...]}."""
-    rows = _parse_tabular_body(tex_path.read_text(), num_data_cols)
-    if not rows:
-        raise RuntimeError(f"No coefficient rows parsed from {tex_path}")
-    return rows
-
-
-def split_tex_row(line: str):
-    r"""Split a tabular row terminated by \\ into cells. Returns None if not a row."""
-    stripped = line.strip()
-    if not stripped.endswith("\\\\"):
-        return None
-    body = stripped[: -len("\\\\")]
-    return body.split("&")
-
-
-def parse_cell(cell: str):
-    """Parse a single coefficient cell. Returns float or None if blank."""
-    trimmed = cell.strip()
-    if trimmed == "" or trimmed == "$$":
-        return None
-    match = COEF_CELL_RE.search(trimmed)
-    return None if match is None else float(match.group(1))
-
-
-def parse_coef_and_stars(cell: str):
-    """Parse a coefficient cell into (value, stars). Returns (None, '') if blank."""
-    trimmed = cell.strip()
-    if trimmed == "" or trimmed == "$$":
-        return (None, "")
-    match = COEF_CELL_RE.search(trimmed)
-    if match is None:
-        return (None, "")
-    return (float(match.group(1)), match.group(2) or "")
-
-
-def parse_se_cell(cell: str):
-    """Parse a standard error cell like $(0.049)$. Returns float or None."""
-    trimmed = cell.strip()
-    if trimmed == "" or trimmed == "$$":
-        return None
-    match = SE_CELL_RE.search(trimmed)
-    return None if match is None else float(match.group(1))
-
-
-# =====
-# .tex parsing: SE rows paired with coefficient rows
-# =====
-def parse_tex_with_details(tex_path: Path, num_data_cols: int) -> dict:
-    """Parse .tex into {label: [(coef, se, stars), ...]} by pairing coef+SE rows."""
-    text = tex_path.read_text()
-    details = {}
-    state = {"label": None, "coefs": None}
-    for cells in _iter_tabular_rows(text):
-        if len(cells) != num_data_cols + 1:
-            continue
-        _consume_detail_row(cells, state, details)
-    if state["label"] is not None:
-        details[state["label"]] = [(c, None, s) for c, s in state["coefs"]]
-    if not details:
-        raise RuntimeError(f"No detail rows parsed from {tex_path}")
-    return details
-
-
-def _consume_detail_row(cells, state, details):
-    """Update state/details for one coef row (label) or SE row (blank label)."""
-    label = cells[0].strip()
-    if not label and state["label"] is None:
-        return
-    if label:
-        if state["label"] is not None:
-            details[state["label"]] = [(c, None, s) for c, s in state["coefs"]]
-        state["label"] = label
-        state["coefs"] = [parse_coef_and_stars(c) for c in cells[1:]]
-    else:
-        ses = [parse_se_cell(c) for c in cells[1:]]
-        details[state["label"]] = [
-            (coef, se, stars)
-            for (coef, stars), se in zip(state["coefs"], ses)
-        ]
-        state["label"] = None
-        state["coefs"] = None
-
-
-# =====
-# .tex parsing: GoF rows
-# =====
-GOF_LABELS = {
-    "Observations", "AR(1) p-value", "AR(2) p-value", "Sargan p-value",
-    "Peer mean pair sum = 0 (p)", "Max peer pair sum = 0 (p)",
-    "Median peer pair sum = 0 (p)", "Min peer pair sum = 0 (p)",
-}
-
-
-def parse_tex_gof_rows(tex_path: Path, num_data_cols: int) -> dict:
-    """Parse GoF rows into {label: [value_per_col]}. Observations are ints."""
-    text = tex_path.read_text()
-    gof = {}
-    for cells in _iter_tabular_rows(text):
-        if len(cells) != num_data_cols + 1:
-            continue
-        label = cells[0].strip()
-        if label not in GOF_LABELS:
-            continue
-        gof[label] = [_parse_gof_value(c, label) for c in cells[1:]]
-    if not gof:
-        raise RuntimeError(f"No GoF rows parsed from {tex_path}")
-    return gof
-
-
-def _parse_gof_value(cell: str, label: str):
-    """Parse a GoF cell. Observations are ints (strip commas), p-values floats."""
-    trimmed = cell.strip()
-    if trimmed == "" or trimmed == "$$":
-        return None
-    match = re.search(r"\$(-?[\d,]+(?:\.\d+)?)\$", trimmed)
-    if match is None:
-        return None
-    raw = match.group(1).replace(",", "")
-    return int(raw) if label == "Observations" else float(raw)
-
-
-# =====
-# Baseline .tex: coefficient magnitude parity with Table DP1 reference
-# =====
-def _tolerance_for(label: str) -> float:
-    """Return the per-coefficient tolerance, using overrides where set."""
-    return TOLERANCE_OVERRIDES.get(label, TOLERANCE)
-
-
-def _check_coef(label, expected, actual, col_idx, model_label):
-    """Return a failure string if actual does not match expected, else None."""
-    tol = _tolerance_for(label)
-    if actual is None:
-        return (f"{label}: blank cell at column {col_idx} ({model_label}); "
-                f"expected {expected:+.3f}")
-    if abs(actual - expected) > tol:
-        return (f"{label}: got {actual:+.3f}, expected {expected:+.3f} "
-                f"(diff {actual - expected:+.3f} exceeds {tol})")
-    return None
-
-
-@pytest.mark.parametrize("model_label,reference", [
-    ("T1_mean",      T1_MEAN_REFERENCE),
-    ("T2_mean",      T2_MEAN_REFERENCE),
-    ("T1_minmedmax", T1_MINMEDMAX_REFERENCE),
-])
-def test_baseline_coefficients_match_reference(baseline_coefs, model_label, reference):
-    """Every reference coefficient matches the parsed .tex value within tolerance."""
-    col_idx = REFERENCE_COLUMN_INDEX[model_label]
-    failures = []
-    for label, expected in reference.items():
-        if label not in baseline_coefs:
-            failures.append(f"{label}: missing from baseline .tex")
-            continue
-        msg = _check_coef(label, expected, baseline_coefs[label][col_idx], col_idx, model_label)
-        if msg:
-            failures.append(msg)
-    assert not failures, (
-        f"{model_label} reference mismatches:\n  " + "\n  ".join(failures)
-    )
-
-
-def test_baseline_t2_minmedmax_matches_stata(baseline_coefs):
-    """T2 min/med/max column 3 matches pinned values (verified vs Stata within 0.01)."""
-    col_idx = REFERENCE_COLUMN_INDEX["T2_minmedmax"]
-    failures = []
-    for label, expected in T2_MINMEDMAX_REFERENCE.items():
-        if label not in baseline_coefs:
-            failures.append(f"{label}: missing from baseline .tex")
-            continue
-        msg = _check_coef(label, expected, baseline_coefs[label][col_idx], col_idx, "T2_minmedmax")
-        if msg:
-            failures.append(msg)
-    assert not failures, "T2_minmedmax mismatches:\n  " + "\n  ".join(failures)
-
-
-def test_baseline_has_four_data_columns(baseline_coefs):
-    """Baseline .tex must expose 4 model columns (T1 mean, T2 mean, T1 mmm, T2 mmm)."""
-    for label, row in baseline_coefs.items():
-        assert len(row) == 4, (
-            f"Row '{label}' has {len(row)} columns, expected 4"
-        )
-
-
-# =====
-# Baseline .tex: standard errors and significance stars
-# =====
-def test_baseline_t1_mean_se_and_stars(baseline_details):
-    """T1 mean SEs and stars match the Stata reference row-by-row."""
-    col_idx = REFERENCE_COLUMN_INDEX["T1_mean"]
-    failures = []
-    for label, expected_se in T1_MEAN_SE_REFERENCE.items():
-        if label not in baseline_details:
-            failures.append(f"{label}: missing from baseline .tex")
-            continue
-        _, se, stars = baseline_details[label][col_idx]
-        if se is None or abs(se - expected_se) > SE_TOLERANCE:
-            failures.append(
-                f"{label} SE: got {se}, expected {expected_se:.3f} "
-                f"(tol {SE_TOLERANCE})"
-            )
-        expected_stars = T1_MEAN_STARS_REFERENCE[label]
-        if stars != expected_stars:
-            failures.append(
-                f"{label} stars: got '{stars}', expected '{expected_stars}'"
-            )
-    assert not failures, "T1_mean SE/stars mismatches:\n  " + "\n  ".join(failures)
-
-
-# =====
-# Baseline .tex: Goodness-of-fit row assertions
-# =====
-def test_baseline_t1_mean_gof_rows(baseline_gof):
-    """T1 mean GoF rows (Obs, AR, Sargan, Wald) match reference and pass diagnostics."""
-    col_idx = REFERENCE_COLUMN_INDEX["T1_mean"]
-    _assert_gof_column(baseline_gof, col_idx, T1_MEAN_GOF_REFERENCE, "T1_mean")
-
-
-def test_baseline_t2_mean_gof_rows(baseline_gof):
-    """T2 mean GoF rows match reference."""
-    col_idx = REFERENCE_COLUMN_INDEX["T2_mean"]
-    _assert_gof_column(baseline_gof, col_idx, T2_MEAN_GOF_REFERENCE, "T2_mean")
-
-
-def test_baseline_t1_minmedmax_gof_rows(baseline_gof):
-    """T1 min/med/max GoF rows match reference."""
-    col_idx = REFERENCE_COLUMN_INDEX["T1_minmedmax"]
-    _assert_gof_column(baseline_gof, col_idx, T1_MINMEDMAX_GOF_REFERENCE, "T1_minmedmax")
-
-
-def test_baseline_diagnostics_pass_thresholds(baseline_gof):
-    """AR(2) and Sargan p-values > 0.05 across all four baseline columns."""
-    failures = []
-    for col_idx in range(4):
-        ar2 = baseline_gof["AR(2) p-value"][col_idx]
-        sargan = baseline_gof["Sargan p-value"][col_idx]
-        if ar2 is None or ar2 <= 0.05:
-            failures.append(f"col {col_idx}: AR(2) p-value {ar2} not > 0.05")
-        if sargan is None or sargan <= 0.05:
-            failures.append(f"col {col_idx}: Sargan p-value {sargan} not > 0.05")
-    assert not failures, "Diagnostic thresholds violated:\n  " + "\n  ".join(failures)
-
-
-def _assert_gof_column(gof, col_idx, reference, model_label):
-    """Validate one model column against the reference GoF dict."""
-    failures = []
-    for label, expected in reference.items():
-        if label not in gof:
-            failures.append(f"{label}: missing from .tex GoF rows")
-            continue
-        actual = gof[label][col_idx]
-        if label == "Observations":
-            if actual != expected:
-                failures.append(f"{label}: got {actual}, expected {expected}")
-        elif actual is None or abs(actual - expected) > GOF_TOLERANCE:
-            failures.append(
-                f"{label}: got {actual}, expected {expected:.3f} "
-                f"(tol {GOF_TOLERANCE})"
-            )
-    assert not failures, f"{model_label} GoF mismatches:\n  " + "\n  ".join(failures)
-
-
-# =====
-# Extended .tex: structural checks for 12 columns
-# =====
 def test_extended_has_twelve_data_columns(extended_coefs):
     """Extended .tex must expose 12 model columns (4 baselines x 3 specs)."""
     for label, row in extended_coefs.items():
         assert len(row) == 12, (
             f"Row '{label}' in extended has {len(row)} columns, expected 12"
-        )
-
-
-def test_extended_col0_matches_baseline_t1_mean(baseline_coefs, extended_coefs):
-    """Extended col 0 (T1 mean Base) should match baseline col 0 within tolerance."""
-    _assert_extended_matches_baseline(baseline_coefs, extended_coefs,
-                                      T1_MEAN_REFERENCE, base_col=0, ext_col=0)
-
-
-def test_extended_col6_matches_baseline_t1_minmedmax(baseline_coefs, extended_coefs):
-    """Extended col 6 (T1 min/med/max Base) should match baseline col 2 within tolerance."""
-    _assert_extended_matches_baseline(baseline_coefs, extended_coefs,
-                                      T1_MINMEDMAX_REFERENCE, base_col=2, ext_col=6)
-
-
-def _assert_extended_matches_baseline(baseline_coefs, extended_coefs, reference,
-                                      base_col, ext_col):
-    """Assert baseline[base_col] == extended[ext_col] for each reference label."""
-    for label in reference:
-        base_val = baseline_coefs[label][base_col]
-        ext_val = extended_coefs[label][ext_col]
-        assert base_val is not None and ext_val is not None, (
-            f"{label}: baseline={base_val}, extended={ext_val} (expected both non-null)"
-        )
-        tol = _tolerance_for(label)
-        assert abs(base_val - ext_val) <= tol, (
-            f"{label}: baseline col {base_col} = {base_val:+.3f} but extended "
-            f"col {ext_col} = {ext_val:+.3f} (diff exceeds {tol})"
         )
 
 
@@ -619,11 +62,9 @@ def test_extended_chat_facial_coefficients_pinned(extended_coefs):
             failures.append(f"{label}: missing from extended .tex")
             continue
         actual = extended_coefs[label][col_idx]
-        msg = _check_coef(label, expected, actual, col_idx, f"extended[{col_idx}]")
+        msg = check_coef(label, expected, actual, col_idx, f"extended[{col_idx}]")
         if msg:
             failures.append(msg)
-    assert not failures, "Extended chat/facial pin mismatches:\n  " + "\n  ".join(failures)
-
-
-if __name__ == "__main__":
-    main()
+    assert not failures, (
+        "Extended chat/facial pin mismatches:\n  " + "\n  ".join(failures)
+    )
