+++ wiki/methods/dynamic-regression.md
@@ -88,7 +88,8 @@ Each of the 4 baseline columns appears 3 times: `{Baseline, +Chat, +Chat+Facial}
 Total: **143 tests passing** across three categories.
 
 - **Panel structure (116 tests):** `tests/test_dynamic_regression_panel.py` (46) + `tests/test_dynamic_regression_merged_panel.py` (51) + `tests/test_dynamic_regression_minmedmax.py` (19). Pins row counts, merge integrity, NaN patterns, lag correctness, deviation roundtrips, min/med/max peer-stat correctness, and hand-verified edge rows (tied min=med, mixed more/less, cross-supergame lag).
-- **Stata parity (14 tests):** `tests/test_dynamic_regression_significance.py`. Verifies baseline-table coefficients (T1 mean, T2 mean, T1 min/med/max, T2 min/med/max), standard errors, significance stars, and GoF rows (Observations=1520, AR(2)>0.05, Sargan>0.05, pairwise Wald p-values) against Stata DP1 reference. Default tolerance **0.005** with per-coefficient overrides: Round 1 at 0.01 (GMM two-step optimizer noise) and `contmoremax_L1` at 0.006 (Stata 3-decimal rounding boundary). Also pins 21 Chat/Facial coefficients in the extended spec and verifies baseline columns equal their counterparts in the extended table.
+- **Stata parity (12 tests):** `tests/test_dynamic_regression_parity.py`. Verifies baseline-table coefficients (T1 mean, T2 mean, T1 min/med/max, T2 min/med/max), standard errors, significance stars, and GoF rows (Observations=1520, AR(2)>0.05, Sargan>0.05, pairwise Wald p-values) against Stata DP1 reference. Default tolerance **0.005** with per-coefficient overrides: Round 1 at 0.01 (GMM two-step optimizer noise) and `contmoremax_L1` at 0.006 (Stata 3-decimal rounding boundary). Also cross-checks that baseline columns equal their counterparts in the extended table.
+- **Extended structure (2 tests):** `tests/test_dynamic_regression_significance.py`. Pins 21 Chat/Facial coefficients across extended cols 1,2,4,5,7,8,10,11 and asserts 12 data columns are present.
 - **Pipeline safety (13 tests):** `tests/test_dynamic_regression_pipeline_safety.py`. Covers `safe_left_merge` (duplicate/missing merge-key failures), `fill_no_message_rounds` NaN-bound guard, and `convert_made_promise` NaN error path.
 
 ## Related
